review feedback, fixing test
