+++ auth/auth_std.go
